requirements.txt.txt
Flask
google-generativeai
markdown
gunicorn
Flask-Cors
